Move core.py contents into core package
Create a core package to separate code for each function to improve each file's readability
