+++ datashelf/core/delete.py
@@ -1,0 +1,2 @@
+def remove():
+    ...
